# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: CalBot API
  version: 0.1.0
  description: API for managing staff authentication and CalBot content.
servers:
  - url: http://localhost:8080
paths:
...
  /public/menu-options/{id}:
    get:
      tags: [Public menu]
      summary: Get an active menu option, one weighted-random active prompt, and active child options
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: The selected option and its available next options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuOptionDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Active menu option not found
  /public/menu-options/{id}/responses:
    get:
      tags: [Public menu]
      summary: Get one weighted-random active response group and its active response items
      description: Each active response group is selected with probability proportional to its weight. The response items inside the selected group remain ordered by sort_order. Returns null when the option has no active response group.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: One selected active response group and its ordered active items, or null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGroup'
                nullable: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Active menu option not found
  /public/content-feedback:
    post:
      tags: [Public feedback]
      summary: Submit a standalone feedback message for future staff review
      description: This endpoint only stores the submitted name, IP address, and message. It never changes menu content automatically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentFeedbackRequest'
      responses:
        '201':
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentFeedbackCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    MenuOption:
      type: object
      required: [id, title, sort_order]
      properties:
        id:
          type: integer
          format: int64
        parent_option_id:
          type: integer
          format: int64
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        sort_order:
          type: integer
    MenuOptionDetail:
      allOf:
        - $ref: '#/components/schemas/MenuOption'
        - type: object
          required: [options]
          properties:
            prompt:
              allOf:
                - $ref: '#/components/schemas/MenuOptionPrompt'
              nullable: true
            options:
              type: array
              description: Active child options ordered by sort order.
              items:
                $ref: '#/components/schemas/MenuOption'
    MenuOptionPrompt:
      type: object
      required: [id, message]
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
    ResponseGroup:
      type: object
      required: [id, name, items]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ResponseItem'
    ResponseItem:
      type: object
      required: [id, type, sort_order]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [TEXT, LINK, IMAGE, YOUTUBE, VIDEO, DOCUMENT]
        text:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        sort_order:
          type: integer
    ContentFeedbackRequest:
      type: object
      required: [full_name, ip, message]
      properties:
        full_name:
          type: string
        ip:
          type: string
          format: ipv4
        message:
          type: string
    ContentFeedbackCreated:
      type: object
      required: [id, created_at]
      properties:
        id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string